pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hbmon"
version = "0.1.0"
description = "LAN hummingbird monitoring: YOLO detection + CLIP species + individual re-id + web UI"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.110,<1",
  "uvicorn[standard]>=0.27,<1",
  "gunicorn>=21,<24",
  "jinja2>=3.1,<4",
  "python-multipart>=0.0.9,<1",
  "SQLAlchemy>=2.0,<3",
  "aiosqlite>=0.20,<1",
  "asyncpg>=0.29,<1",
  "psycopg[binary]>=3.1,<4",
  "redis>=5,<8",
  "opencv-python-headless>=4.8,<5",
  "numpy>=1.26,<2.0",


  # ML stack
  "scikit-learn>=1.4,<2",
  "ultralytics>=8.3.0,<9",
  "openvino>=2024.0,<2025",
  "torch>=2.2,<3",
  "torchvision>=0.17,<1",
  "open-clip-torch>=2.20,<4",
  "pillow>=10,<13",

  # Background job processing
  "rq>=1.16,<2",
  "rq-scheduler>=0.13,<1",

  # Annotation utils
  "sahi>=0.11.16",
]

# MOVED FROM [project.optional-dependencies] to [dependency-groups]
[dependency-groups]
dev = [
  "pytest>=8,<10",
  "pytest-xdist>=3.5,<4",
  "pytest-playwright>=0.4.4,<1",
  "ruff>=0.6,<1",
  "pytest-cov>=4,<8",
  "pre-commit>=3.6,<5",
  "pytest-sugar>=1.1.1",
  "pyyaml>=6.0,<7",
  "pytest-asyncio>=0.23,<1",
  "pytest-timeout>=2.2,<3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 110
target-version = "py312"
lint.ignore = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require ML dependencies and test data)",
    "unit: marks tests as unit tests (lightweight, no ML dependencies)",
    "ui: marks tests as UI tests (require playwright)",
    "heavy: marks tests as resource-intensive (may crash systems with limited RAM/GPU)",
]
# Heavy tests are excluded by default. Use -m 'heavy' to explicitly run them.

[tool.coverage.report]
show_missing = true
skip_covered = false

asyncio_mode = "auto"
